+++ core/interfaces.go
@@ -44,6 +44,8 @@ type Administration interface {
 	AdmGetAccount(accountID string) (*model.Account, error)
 	AdmGrantAccountRoles(accountID string, appID string, roleIDs []string, l *logs.Log) error
 
+	AdmGrantAccountPermissions(appID string, orgID string, accountID string, permissionNames []string, assignerPermissions []string, l *logs.Log) error
+
 	AdmGetApplicationLoginSessions(appID string, orgID string, identifier *string, accountAuthTypeIdentifier *string,
 		appTypeID *string, appTypeIdentifier *string, anonymoys *bool, deviceID *string, ipAddress *string) ([]model.LoginSession, error)
 	AdmDeleteApplicationLoginSession(appID string, orgID string, currentAccountID string, identifier string, sessionID string, l *logs.Log) error