+++ driver/web/docs/gen/gen_types.go
@@ -537,6 +537,12 @@ type AdminReqCreateApplicationRole struct {
 	Permissions []string `json:"permissions"`
 }
 
+// AdminReqGrantPermissions defines model for _admin_req_grant-permissions.
+type AdminReqGrantPermissions struct {
+	AccountId   string   `json:"account_id"`
+	Permissions []string `json:"permissions"`
+}
+
 // ServicesReqAccountExists defines model for _services_req_account-exists.
 type ServicesReqAccountExists struct {
 	ApiKey            string                           `json:"api_key"`
@@ -1004,6 +1010,9 @@ type PostAdminAccountMfaParams struct {
 // PostAdminAccountRolesJSONBody defines parameters for PostAdminAccountRoles.
 type PostAdminAccountRolesJSONBody AdminReqCreateApplicationRole
 
+// PutAdminApplicationAccountPermissionsGrantJSONBody defines parameters for PutAdminApplicationAccountPermissionsGrant.
+type PutAdminApplicationAccountPermissionsGrantJSONBody AdminReqGrantPermissions
+
 // PutAdminApplicationAccountRolesGrantJSONBody defines parameters for PutAdminApplicationAccountRolesGrant.
 type PutAdminApplicationAccountRolesGrantJSONBody AdminReqAccountRoles
 
@@ -1259,6 +1268,9 @@ type GetTpsServiceRegsParams struct {
 
 // PostAdminAccountRolesJSONRequestBody defines body for PostAdminAccountRoles for application/json ContentType.
 type PostAdminAccountRolesJSONRequestBody PostAdminAccountRolesJSONBody
+
+// PutAdminApplicationAccountPermissionsGrantJSONRequestBody defines body for PutAdminApplicationAccountPermissionsGrant for application/json ContentType.
+type PutAdminApplicationAccountPermissionsGrantJSONRequestBody PutAdminApplicationAccountPermissionsGrantJSONBody
 
 // PutAdminApplicationAccountRolesGrantJSONRequestBody defines body for PutAdminApplicationAccountRolesGrant for application/json ContentType.
 type PutAdminApplicationAccountRolesGrantJSONRequestBody PutAdminApplicationAccountRolesGrantJSONBody